Implemented super simple auth

+++ app/lib/actions.ts
@@ -1,11 +1,16 @@
 "use server";
 import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
 import { db } from "@/db";
 import { NewSong, songs } from "@/db/schema";
 import { eq } from "drizzle-orm";
+import { signIn, signOut, isAuthenticated  } from "@/auth";
 
-export async function insertSong(formData: FormData, lyrics: string) {
+export async function insertSongAction(formData: FormData, lyrics: string) {
+	const authenticated = await isAuthenticated();
+	if (!authenticated) {
+		throw new Error('Unauthorized access: User does not have admin privileges.');
+	}
+
 	const song: NewSong = {
 		title: formData.get("title") as string,
 		melody: formData.get("melody") as string,
@@ -15,11 +20,23 @@ export async function insertSong(formData: FormData, lyrics: string) {
 
 	revalidatePath("/songs");
 }
-export async function deleteSong(id: number) {
+export async function deleteSongAction(id: number) {
+	const authenticated = await isAuthenticated();
+	if (!authenticated) {
+		throw new Error('Unauthorized access: User does not have admin privileges.');
+	}
+
+	console.log("deleteSongAction", id, authenticated);
+
 	await db.delete(songs).where(eq(songs.id, id));
 	revalidatePath("/songs");
 }
-export async function updateSong(id: number, formData: FormData, lyrics: string) {
+export async function updateSongAction(id: number, formData: FormData, lyrics: string) {
+	const authenticated = await isAuthenticated();
+	if (!authenticated) {
+		throw new Error('Unauthorized access: User does not have admin privileges.');
+	}
+
 	const song: NewSong = {
 		title: formData.get("title") as string,
 		melody: formData.get("melody") as string,
@@ -28,4 +45,14 @@ export async function updateSong(id: number, formData: FormData, lyrics: string)
 	await db.update(songs).set(song).where(eq(songs.id, id));
 
 	revalidatePath("/songs");
+}
+
+
+export async function authenticateAction(formData: FormData) {
+	const password = formData.get("password") as string;
+	return signIn(password);
+  }
+
+export async function signOutAction() {
+	signOut();
 }
